pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gradients-worker"
version = "0.1.0"
authors = [{ name = "besimray", email = "besimray24@gmail.com" }]
description = "Worker for training and evaluating models using Gradients API"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.28.0",
    "python-dotenv>=0.19.0",
    "wandb>=0.13.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "pydantic-settings>=0.10.0",
    "lm_eval[wandb] @ git+https://github.com/besimray/lm-evaluation-harness.git@pubmedqa-use-meshes",
    "peft>=0.5.0",
    "huggingface-hub>=0.19.0",
    "transformers>=4.37.0",
    "torch>=2.0.0",
    "asyncio>=3.4.3",
    "aiohttp-retry>=2.8.3",
    "minio>=7.0.0",
    "datasets>=2.14.0",
    "accelerate>=0.25.0",
    "runpod>=1.5.0",
    "tenacity>=8.0.0"
]

[tool.hatch.build.targets.wheel]
packages = ["src/gradients_worker"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build]
packages = ["src"]
